🚚 test(graph_utils): move tests
