+++ tests/backend/test_graphs.py
@@ -1,212 +1,13 @@
-from collections import defaultdict
-from collections.abc import Sequence
 from typing import Optional
 
 import pytest
 from paicorelib import HwConfig
 
 import paibox as pb
-from paibox.backend.graphs import get_node_degrees, get_pred_dg_by_succ_dg, toposort
+from paibox.backend.graph_utils import *
 from paibox.backend.types import *
 from paibox.components import Neuron
 from paibox.exceptions import GraphBuildError, GraphConnectionError, NotSupportedError
-
-from .conftest import TestData
-
-
-class TestTopoSort:
-    @pytest.mark.parametrize(
-        TestData.toposort_data["args"],
-        TestData.toposort_data["data"],
-        ids=TestData.toposort_data["ids"],  # type:ignore
-    )
-    def test_toposort(self, nodes):
-        """
-        Test #1: one input 1
-            INP1 -> N1 -> N2 -> N3
-            N1 -> N4
-            N4 -> N2
-
-        Test #2: one input 2
-            INP1 -> N1 -> N2 -> N3 -> N4
-            N1 -> N5 -> N6 -> N7 -> N4
-            N5 -> N3
-            N3 -> N6
-
-        Test #3: more than one input
-            INP1 -> N1 -> N2 -> N3
-            INP2 -> N4 -> N5 -> N3
-            N2 -> N4
-
-        Test #4: one input 3
-            INP1 -> N1 -> N2 -> N4
-            N1 -> N3 -> N4
-
-        Test #5: headless neuron 1
-            INP1 -> N1 -> N2 -> N4
-            N3 -> N2
-        """
-        ordered = toposort(nodes)
-        assert len(ordered) == len(nodes)
-
-    @pytest.mark.parametrize(
-        "edges",
-        [
-            {
-                "inp1": {"n1"},
-                "n1": {"n2"},
-                "n2": {"n3", "n4"},
-                "n3": {},
-                "n4": {"n1"},
-            },
-            {"inp1": {"n1"}, "n1": {"n2"}, "n2": {"n3"}, "n3": {"n2", "n4"}, "n4": {}},
-            {"inp1": {"n1"}, "inp2": {"n2"}, "n1": {"n3"}, "n2": {"n1"}, "n3": {"n2"}},
-            {
-                "inp1": {"n1"},
-                "inp2": {"n2"},
-                "inp3": {"n2"},
-                "n1": {"n3"},
-                "n2": {"n1"},
-                "n3": {"n4"},
-                "n4": {"n2"},
-            },
-        ],
-    )
-    def test_toposort_has_cycle(self, edges):
-        """
-        Test #1: 1 input
-        INP1 -> N1 -> N2 -> N3
-        N4 -> N1
-        N2 -> N4
-
-        Test #2: 1 input
-        INP1 -> N1 -> N2 -> N3 -> N4
-        N3 -> N2
-
-        Test #3: 2 inputs
-        INP1 -> N1 -> N3 -> N2
-        INP2 -> N2
-
-        Test #4: 3 inputs
-        INP1 -> N1 -> N3 -> N4 -> N2
-        INP2 -> N2 -> N1
-        INP3 -> N2
-        """
-        with pytest.raises(NotSupportedError):
-            ordered = toposort(edges)
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.parametrize(
-    "constrs, expected",
-    [
-        ([{"1", "2", "3"}, {"4", "5"}], ["1", "2", "3"]),
-        ([{"1", "2"}, {"3", "4"}, {"5", "6"}], ["1", "2"]),
-    ],
-)
-def test_bounded_by(constrs, expected):
-    def _bounded_by_proto(node: str, constrs: Sequence[Sequence[str]]) -> list[str]:
-        for constr in constrs:
-            for bounded_node in constr:
-                if node == bounded_node:
-                    return list(set(constr))
-
-        return []
-
-    result = _bounded_by_proto("1", constrs)
-
-    assert set(result) == set(expected)
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.parametrize(
-    "constrs, expected",
-    [
-        ({"1": {"2", "3", "4"}, "2": {"3", "4"}}, ["2", "3", "4"]),
-        ({"1": {"2", "3"}, "4": {"1"}}, ["2", "3", "4"]),
-        ({"2": {"1", "3"}, "4": {"1"}}, ["2", "4"]),
-    ],
-)
-def test_conflicted_by(constrs, expected):
-    def _conflicted_by_proto(node: str, constrs: dict[str, Sequence[str]]) -> list[str]:
-        c = set(constrs.get(node, []))
-
-        for k, v in constrs.items():
-            for conf_node in v:
-                if node == conf_node:
-                    c.add(k)
-
-        return list(c)
-
-    result = _conflicted_by_proto("1", constrs)
-
-    assert set(result) == set(expected)
-
-
-@pytest.mark.skip(reason="Not implemented")
-@pytest.mark.parametrize(
-    "constrs, expected",
-    [
-        (
-            [{"1", "2", "3"}, {"4", "5"}, {"6", "7", "1"}, {"4", "8"}],
-            [frozenset(["1", "2", "3", "6", "7"]), frozenset(["4", "5", "8"])],
-        ),
-        (
-            [{"1", "2", "3"}, {"4", "5"}, {"6", "7"}, {"4", "8"}],
-            [
-                frozenset(["1", "2", "3"]),
-                frozenset(["6", "7"]),
-                frozenset(["4", "5", "8"]),
-            ],
-        ),
-        (
-            [{"1", "2", "3"}, {"4", "5"}, {"6", "7"}, {"4", "1", "8"}],
-            [
-                frozenset(["1", "2", "3", "4", "5", "8"]),
-                frozenset(["6", "7"]),
-            ],
-        ),
-    ],
-)
-def test_bounded_nodes_check(constrs, expected):
-    def _bounded_nodes_check_proto(
-        constrs: list[Sequence[str]],
-    ) -> list[frozenset[str]]:
-        seen = {}
-        need_update_nodes = []
-
-        for bounded_set in constrs:
-            for node in bounded_set:
-                if node in seen:
-                    need_update_nodes.append(node)
-                    seen[node].update(bounded_set)
-                else:
-                    seen[node] = bounded_set
-
-        _constr = []
-
-        for bounded in constrs:
-            for node in need_update_nodes:
-                flag = True
-                if node in bounded:
-                    flag = False
-                    break
-
-            if flag:
-                # Unique set
-                _constr.append(frozenset(bounded))
-
-        while need_update_nodes:
-            for node in need_update_nodes:
-                _constr.append(frozenset(seen[node]))
-                need_update_nodes.remove(node)
-
-        return _constr
-
-    result = _bounded_nodes_check_proto(constrs)
-
-    # Inconvenient to compare the elements in the list of sets.
-    assert len(result) == len(expected)
 
 
 class TestPAIGraph:
@@ -502,117 +303,3 @@ class TestGroupEdges:
                 break
 
         assert pos_n2 != pos_n3
-
-
-class TestDAGPathDistance:
-    @staticmethod
-    def get_longest_path_proto(
-        edges_with_d: dict[NodeName, dict[NodeName, int]], ordered_nodes: list[NodeName]
-    ) -> tuple[list[NodeName], int]:
-        """Get the longest path in the DAG.
-
-        Args:
-            - edges_with_d: a list of directed edges with distance.
-            - ordered_nodes: nodes in topological sorting.
-
-        Return: the longest distance in the graph.
-        """
-        distances: dict[NodeName, int] = {node: 0 for node in ordered_nodes}
-        pred_nodes: dict[NodeName, NodeName] = defaultdict()
-
-        for node in ordered_nodes:
-            for neighbor in edges_with_d[node]:
-                d = edges_with_d[node][neighbor]
-                if distances[node] + d > distances[neighbor]:
-                    distances[neighbor] = distances[node] + d
-                    pred_nodes[neighbor] = node
-
-        # When there are more than one output nodes
-        # with same distance, choose the first one.
-        node = max(
-            filter(lambda node: len(edges_with_d[node]) == 0, distances),
-            key=lambda node: distances.get(node, 0),
-        )
-
-        distance = distances[node]
-        path = [node]
-
-        while path[-1] in pred_nodes:
-            path.append(pred_nodes[path[-1]])
-
-        path.reverse()
-        return path, distance
-
-    @pytest.mark.parametrize(
-        TestData.get_longest_path_data["args"],
-        TestData.get_longest_path_data["data"],
-        ids=TestData.get_longest_path_data["ids"],  # type:ignore
-    )
-    def test_get_longest_path_proto(self, edges, expected_path, expected_distance):
-        ordered = toposort(edges)
-        path, distance = self.get_longest_path_proto(edges, ordered)
-
-        assert path == expected_path
-        assert distance == expected_distance
-
-    @staticmethod
-    def get_shortest_path_proto(
-        edges_with_d: dict[NodeName, dict[NodeName, int]],
-        ordered_nodes: list[NodeName],
-        input_nodes: list[NodeName],
-    ) -> tuple[list[NodeName], int]:
-        """Get the shortest path in the DAG.
-
-        Args:
-            - edges_with_d: a list of directed edges with distance.
-            - ordered_nodes: nodes in topological sorting.
-            - input_nodes: input nodes.
-
-        Return: the shortest distance in the graph.
-        """
-        distances: dict[NodeName, int] = defaultdict(lambda: 999)
-        pred_nodes: dict[NodeName, NodeName] = defaultdict()
-
-        # set initial value for all inputs nodes.
-        if input_nodes:
-            for inode in input_nodes:
-                distances[inode] = 0
-        else:
-            distances[ordered_nodes[0]] = 0
-
-        for node in ordered_nodes:
-            for neighbor in edges_with_d[node]:
-                d = edges_with_d[node][neighbor]
-                if distances[node] + d < distances[neighbor]:
-                    distances[neighbor] = distances[node] + d
-                    pred_nodes[neighbor] = node
-
-        # When there are more than one output nodes
-        # with same distance, choose the first one.
-        node = min(
-            filter(lambda node: len(edges_with_d[node]) == 0, distances),
-            key=lambda node: distances.get(node, 0),
-        )
-
-        distance = distances[node]
-        path = [node]
-
-        while path[-1] in pred_nodes:
-            path.append(pred_nodes[path[-1]])
-
-        path.reverse()
-        return path, distance
-
-    @pytest.mark.parametrize(
-        TestData.get_shortest_path_data["args"],
-        TestData.get_shortest_path_data["data"],
-        ids=TestData.get_shortest_path_data["ids"],  # type:ignore
-    )
-    def test_get_shortest_path_proto(
-        self, edges, inodes, expected_path, expected_distance
-    ):
-        ordered = toposort(edges)
-        path, dist = self.get_shortest_path_proto(edges, ordered, inodes)
-
-        assert path == expected_path
-        assert dist == expected_distance
